implementing creation and deletion in Employee class

+++ src/main/java/com/chichos_snack_project/service/Delete_employee.java
@@ -11,12 +11,18 @@ public class Delete_employee {
     private static final EmployeeDAPImpl employeeDAO = new EmployeeDAPImpl(AppConfig.getDatasource());
     private static final Logger log = LogManager.getLogger(Delete_employee.class);
 
-    public static boolean delete(int id){
-        try{
-           employeeDAO.delete(id);
-           return true;
-        }catch (SQLException e){
-            log.error("No se pudo eliminar al empleado por una SQLException en el metodo delete de EmployeeDAOImpl");
+    public static boolean delete(int id,String code,String code_entered){
+        if(code.equals(code_entered)){
+            try{
+                log.info("Deleting employee with id: " + id);
+                employeeDAO.delete(id);
+                return true;
+            }catch (SQLException e){
+                log.error("No se pudo eliminar al empleado por una SQLException en el metodo delete de EmployeeDAOImpl "+e.getSQLState());
+                return false;
+            }
+        }else{
+            log.info("El codigo ingresado por el usuario no coincide");
             return false;
         }
     }
